refactor(api): 공통 스키마 파일 폴더 이전
