+++ packages/api/src/index.ts
@@ -6,5 +6,5 @@ export * from "./errors/APIError";
 export * from "./rest/index";
 export * from "./rest/rest";
 export * from "./rest/types";
-export * from "./schema";
+export * from "./schemas";
 export * from "./services";
